Fix inspection of methods accepting events vs. rejections
